requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow==10.2.0
httpx==0.25.0
python-dotenv==1.0.1
pydantic==2.7.1
supabase==2.5.0
google-generativeai==0.5.2